infra/drm/requirements.txt
Flask==3.0.3
PyJWT==2.8.0
prometheus_client==0.20.0
